python-app/requirements.txt
# python-app requirements
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
